feat: harden cluster Raft snapshots and correctness
- enforce Raft current-term commit semantics with leader NOOP entries, conflict-aware AppendEntries match hints, and joint-consensus election quorums

- fail startup on corrupt Raft hard state instead of resetting persisted term and vote data

- stream Raft snapshot installs through CRC32C-protected staging without buffering full snapshots in memory

- include committed and joint membership metadata in InstallSnapshot and AKRL5 log headers so membership survives compaction and restart

- make snapshot completion WAL-transactional and recover only committed snapshot batches

- require pinned peer identities for secure Raft and replication transports while keeping PLAIN explicit for local unauthenticated tests

- harden Raft RPC handling with bounded active handlers, close-time draining, exception containment, and defensive parser validation

- expand cluster smoke coverage for deterministic Raft leadership, large Blob snapshot catch-up, leader churn, hard-state corruption, membership restart-after-compaction, secure pin enforcement, and snapshot staging CRC rejection

- update cluster spec/progress documentation and build revision

+++ akkara/akkengine/include/akk/engine/wal/WalWriter.hpp
@@ -100,7 +100,7 @@ namespace akkaradb::engine::wal {
                 std::span<const uint8_t> key,
                 std::span<const uint8_t> value,
                 uint64_t seq,
-                uint8_t flags,
+                uint16_t flags,
                 uint64_t precomputedFp64 = 0,
                 WalAppendAck ack = WalAppendAck::WRITTEN
             );
